add build

+++ api-dashboard/src/components/Table/OrderDataTable.component.js
@@ -10,7 +10,7 @@ import Loader from '../Loader/Loader.component';
 import Colors from '../../utils/constants/colors.constant';
 import {editOrder, removeOrder} from '../../redux/actions/orders.action';
 
-function OrderDataTable({dataKeys, showActiveOrders=false, providerId}) {
+function OrderDataTable({dataKeys, showInProgress=false, providerId}) {
     const dispatch = useDispatch();
     const currentUserRole = useSelector(({AuthReducer}) => AuthReducer.userRole);
     const [, ordersArray] = useSelectorAsAnArray(({OrdersReducer}) => OrdersReducer.orders);
@@ -30,50 +30,56 @@ function OrderDataTable({dataKeys, showActiveOrders=false, providerId}) {
             <TableHead>
                 <tr>
                     {dataKeys.map( key => <th key={uniqid()}> {formatHeadline(key)} </th>)}
-                    {!showActiveOrders && <th> Actions </th>}
+                    {!showInProgress && <th> Actions </th>}
                 </tr>
             </TableHead>
             <TableBody>
-                {ordersArray.map( order => order.isOrderActive === showActiveOrders &&
-                    <tr>
-                        {dataKeys.map( key => {
-                            const item = order[key];
+                {ordersArray.map( order => {
 
-                            if(Array.isArray(item))
-                                return ( 
-                                    <td key={uniqid()}>
-                                        {item.map((i,index) => {
-                                            if(typeof i === 'object')
-                                                return (
-                                                    <Link key={uniqid()} onClick={() => handleLinkClick(i)}>
-                                                        {`${formatHeadline(key)}${index + 1},`}
-                                                    </Link>
-                                                );
-                                            return <p>{i}</p>;
-                                        })}
-                                    </td> 
-                                );
-                            
-                            if(typeof item === 'object')
-                                return (
-                                    <td key={uniqid()}>
-                                        <Link onClick={() => handleLinkClick(item)}>{formatHeadline(key)}</Link>
-                                    </td> 
-                                );
-                            
-                            if(typeof item === 'boolean')
-                                return (<td key={uniqid()}> {item ? '✔' : '❌'} </td>);
+                    if(showInProgress && (!order.providerId || order.providerId === '-')) return null;
+                    if(!showInProgress && order.providerId && order.providerId === '-') return null;
 
-                            return (<td key={uniqid()}> {formatText(item)} </td>);
-                        })}
-                        {!showActiveOrders && <td>
-                            {currentUserRole === 'ADMIN' ? 
-                                <span onClick={() => deleteItem(order._id)}> <DeleteIcon /> </span> :
-                                <span onClick={() => assignItemToProvider(order._id)}> <TruckIcon /> </span> 
-                            }
-                        </td>}
-                    </tr>
-                )}
+                    return(
+                        <tr>
+                            {dataKeys.map( key => {
+                                const item = order[key];
+
+                                if(Array.isArray(item))
+                                    return ( 
+                                        <td key={uniqid()}>
+                                            {item.map((i,index) => {
+                                                if(typeof i === 'object')
+                                                    return (
+                                                        <Link key={uniqid()} onClick={() => handleLinkClick(i)}>
+                                                            {`${formatHeadline(key)}${index + 1},`}
+                                                        </Link>
+                                                    );
+                                                return <p>{i}</p>;
+                                            })}
+                                        </td> 
+                                    );
+                                
+                                if(typeof item === 'object')
+                                    return (
+                                        <td key={uniqid()}>
+                                            <Link onClick={() => handleLinkClick(item)}>{formatHeadline(key)}</Link>
+                                        </td> 
+                                    );
+                                
+                                if(typeof item === 'boolean')
+                                    return (<td key={uniqid()}> {item ? '✔' : '❌'} </td>);
+
+                                return (<td key={uniqid()}> {formatText(item)} </td>);
+                            })}
+                            {!showInProgress && <td>
+                                {currentUserRole === 'ADMIN' ? 
+                                    <span onClick={() => deleteItem(order._id)}> <DeleteIcon /> </span> :
+                                    <span onClick={() => assignItemToProvider(order._id)}> <TruckIcon /> </span> 
+                                }
+                            </td>}
+                        </tr>
+                    );
+                })}
             </TableBody>
             <PopUpWrapper display={isPopUpOn}>
                 <PopUp display={isPopUpOn}>
